Moved shared code to new files; Added copyright headers;
