+++ com.kf.mvvm-data-binding/Runtime/Theming/ThemeValueList.cs
@@ -10,121 +10,6 @@ using UnityEngine.Events;
 
 namespace MVVMDatabinding.Theming
 {
-    public enum ThemeVariant
-    {
-        Light,
-        Dark,
-        HighContrast,
-    }
-
-    public interface IThemeValue
-    {
-        Type DataType { get; }
-
-        UnityEvent ValueChanged { get; }
-
-        bool ExcludedFromVariants { get; set; }
-
-        void SetVariant(ThemeVariant variant);
-
-#if UNITY_EDITOR
-        // NOTE: We're only using object here because it's strictly editor-only code.
-        object Editor_GetValue();
-#endif
-    }
-
-    [Serializable]
-    public abstract class ThemeValue : IThemeValue
-    {
-        protected ThemeVariant activeVariant { get; set; } = ThemeVariant.Light;
-        public abstract Type DataType { get; }
-        public UnityEvent ValueChanged { get; protected set; } = new UnityEvent();
-
-        [HideInInspector]
-        [SerializeField]
-        protected bool excludedFromVariants = false;
-
-
-        public bool ExcludedFromVariants
-        {
-            get => excludedFromVariants;
-            set
-            {
-                excludedFromVariants = value;
-            }
-        }
-
-        public void SetVariant(ThemeVariant variant)
-        {
-            activeVariant = variant;
-            ValueChanged?.Invoke();
-        }
-
-#if UNITY_EDITOR
-        public abstract object Editor_GetValue();
-#endif
-    }
-
-    public abstract class ThemeValue<T> : ThemeValue
-    {
-        [ConditionalVisibility(nameof(ExcludedFromVariants), ConditionResultType.ShowIfNotEquals)]
-        [SerializeField]
-        private T light;
-
-        [ConditionalVisibility(nameof(ExcludedFromVariants), ConditionResultType.ShowIfNotEquals)]
-        [SerializeField]
-        private T dark;
-
-        [ConditionalVisibility(nameof(ExcludedFromVariants), ConditionResultType.ShowIfNotEquals)]
-        [SerializeField]
-        private T highContrast;
-
-        [ConditionalVisibility(nameof(ExcludedFromVariants), ConditionResultType.ShowIfEquals)]
-        [SerializeField]
-        private T value;
-
-        public T Value
-        {
-            get
-            {
-                if (ExcludedFromVariants)
-                {
-                    return value;
-                }
-
-                switch (activeVariant)
-                {
-                    case ThemeVariant.Light:
-                        return light;
-                    case ThemeVariant.Dark:
-                        return dark;
-                    case ThemeVariant.HighContrast:
-                        return highContrast;
-                }
-
-                return default(T);
-            }
-        }
-
-        public override Type DataType => typeof(T);
-
-#if UNITY_EDITOR
-        public override object Editor_GetValue()
-        {
-            return excludedFromVariants ? value : light;
-        }
-#endif
-    }
-
-    public class ColorThemeValue : ThemeValue<Color> { }
-    public class GradientThemeValue : ThemeValue<Gradient> { }
-    public class MaterialThemeValue : ThemeValue<Material> { }
-    public class TextureThemeValue : ThemeValue<Texture> { }
-    public class IntThemeValue : ThemeValue<int> { }
-    public class FloatThemeValue : ThemeValue<float> { }
-    public class Vector4ThemeValue : ThemeValue<Vector4> { }
-    public class TMPGradientThemeValue : ThemeValue<TMP_ColorGradient> { }
-    public class FontSettingsThemeValue : ThemeValue<ThemeFontSettings> { }
 
     [Serializable]
     public class ThemeItemValue
